requirement.txt
numpy==1.22.1
scikit-learn==1.0.2
python-dotenv==0.19.2